am243x: eclipse-threadx: Fix various MISRA warnings.
- Fix for various MISRA warnings in the ThreadX specific source code.

Fixes: SITSW-5614

+++ source/kernel/threadx/dpl/common/HeapP_threadx.c
@@ -44,7 +44,7 @@ void   HeapP_construct( HeapP_Object *heap, void *heapAddr, size_t heapSize )
 
     p_pool = (TX_BYTE_POOL *)heap;
 
-    (void)tx_byte_pool_create(p_pool, "dpl pool", heapAddr, heapSize);
+    (void)tx_byte_pool_create(p_pool, "dpl pool", heapAddr, (ULONG)heapSize);
 
 }
 
@@ -64,7 +64,7 @@ void  *HeapP_alloc( HeapP_Object *heap, size_t allocSize )
 
     p_pool = (TX_BYTE_POOL *)heap;
 
-    (void)tx_byte_allocate(p_pool, &ptr, allocSize, TX_NO_WAIT);
+    (void)tx_byte_allocate(p_pool, &ptr, (ULONG)allocSize, TX_NO_WAIT);
 
     return ptr;
 }
