+++ source/kernel/threadx/dpl/common/TaskP_threadx.c
@@ -122,7 +122,7 @@ static uint32_t TaskP_calcCpuLoad(uint64_t taskTime, uint64_t totalTime)
 {
     uint32_t cpuLoad;
 
-    cpuLoad = (taskTime * TaskP_LOAD_CPU_LOAD_SCALE) / totalTime;
+    cpuLoad = (uint32_t)(taskTime * TaskP_LOAD_CPU_LOAD_SCALE) / totalTime;
     if( cpuLoad > TaskP_LOAD_CPU_LOAD_SCALE)
     {
         cpuLoad = TaskP_LOAD_CPU_LOAD_SCALE;
